[AltIndexer] Add ingestion client trait
